requirements.txt
fastapi==0.104.1
pydantic==2.5.0
uvicorn[standard]==0.24.0
supabase==2.15.3
python-dotenv==1.0.0
